requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
discord.py==2.4.0
llama-cpp-python==0.3.1
huggingface-hub==0.25.1
python-dotenv==1.0.1
